backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backbone for the project, providing essential APIs and services."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "cffi==1.17.1",
    "colorama==0.4.6",
    "cryptography==42.0.8",
    "dj-database-url==2.2.0",
    "dj-rest-auth==7.0.1",
    "django==5.0.6",
    "django-allauth==65.9.0",
    "django-cors-headers==4.4.0",
    "django-environ>=0.12.0",
    "django-extensions==3.2.3",
    "django-filter==25.1",
    "djangorestframework==3.15.1",
    "djangorestframework-simplejwt==5.5.0",
    "gunicorn==22.0.0",
    "iniconfig==2.1.0",
    "markupsafe==3.0.2",
    "packaging==25.0",
    "pillow==10.3.0",
    "pluggy==1.6.0",
    "psycopg2-binary==2.9.9",
    "pycparser==2.22",
    "pygments==2.19.2",
    "pyjwt==2.9.0",
    "pyopenssl==24.1.0",
    "pytest==8.4.1",
    "pytest-django==4.11.1",
    "python-dotenv==1.1.1",
    "sqlparse==0.5.3",
    "typing-extensions==4.14.1",
    "tzdata==2024.1",
    "werkzeug==3.1.3",
    "whitenoise==6.7.0",
]

[project.optional-dependencies]
dev = [
    "uv",
    "pip"
]